created new component addStockModal, and made it possible to add stock to portfolio
